Update exercise2Task4-4.yaml
changed the image to python:3.9 because a new bug in the DataOps Toolkit 2.40.1 Kubeflow pipeline test because of reference to the python:3 container image, which has a dependency with PyYAML 5.41 and is trying to use Cython 3.0.  This results in this issue:  https://github.com/yaml/pyyaml/issues/724.  This occurs in the "dataset-snapshot" step  and the "model-snapshot" step of the pipeline example you provide here: https://github.com/NetApp/netapp-dataops-toolkit/blob/main/netapp_dataops_k8s/Examples/Kubeflow/Pipelines/ai-training-run.py.